Improve handling of Redis shards failing
These changes prevent synapse totally locking up when any Redis cache
shard is unavailable. Warnings are logged where Redis is not working
and deletes continue to be retried in the background to ensure proper
invalidation during any outages.

This still relies on the synapse process itself not being rolled, the
assumption here being that we won't roll synapse while a node (w/Redis)
is in a broken state.